refactor: extract config io helpers
